MCP_Server/requirements.txt
# TIA Portal MCP Server Requirements

# Core MCP package for Model Context Protocol
mcp>=1.12.3

# TIA Portal dependencies (Windows-specific)
pythonnet>=3.0.1

# Configuration and data handling
pyyaml>=6.0
pydantic>=2.8.0
pydantic-settings>=2.5.2

# File format support
openpyxl>=3.1.0

# Async and HTTP support (typically satisfied by MCP)
anyio>=4.5
httpx>=0.27
jsonschema>=4.20.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0